Return response after update
Right now it's impossible to do anything with the response from onUpdate() when subclassing. I'd like to be able to examine the response object and take further actions depending on the result.

To make this possible, I've returned the response, which I think should make it possible to use it if needed.